driver list: whip up stupid application which enumerates drivers
Create an autotoolized project which builds an application that lists
drivers which are kept in a library's special code sections. A linker
script provides support to collect all drivers' registration details,
and to iterate over that list, including termination.

This is untested, and is an intermediate step before creating more
complex library setups.