driver list: move iteration out of app and into library
Introduce an intermediate library between the driver implementation and
the application. This library is supposed to collect all the driver
registration details and provide the iteration logic, so that apps need
not deal with internals of the library.

Increase verbosity in the construction of that library, and try to
override the _AR spec. Which does not work here (won't take effect,
always ends up running "ar cru" instead of "ld -r" or the "echo" that is
used here for demonstration).