driver list: avoid unexpected NULL base address, improve extern decl
Declaring a pointer seems to also create the variable and initialize its
value to zero. Delcaring an external array instead does result in the
absence of the variable "in code", and does reference the linker
provided symbol instead, which is the address of a memory range.

Print the start and end address of the drivers list before iteration. To
spot when the list is "terminated at its beginning" although more data
is available.