driver list: experiment with intermediate shared object
The previous implementation used noinst_ which translates to archives
(.a or .la). Add another intermediate step of a shared library (.so
installed in lib_) and an executable which uses this shared object. This
is closer to the libsigrok setup.

Still doesn't work. The libtool utility is sooo helpful and rearranges
the order of command line arguments, breaking the whole archive spec for
the drivers library. Also experiments with _LINK since _AR would not
take effect in previous attempts.

Also here (did not bother to separate): git ignore autoscan files.