driver list: move driver list iteration out of main application code
Put the iteration code for the runtime constructor approach out of the
main application and into the driver list handling compile unit, where
it should have gone in the first place. Eliminate the unnecessary extra
source file, which is a (compile time chosen) alternative to the driver
list handling code that uses the linker section. See 'git diff HEAD~2'
for the essence of the runtime constructor approach.