build: use --no-whole-archive for autotools instead of --push-state
The --push-state linker option isn't universally available, so let's do
an explicit --no-whole-archive after our library.

It's unfortunate that this changes the "global state" of the linker
while it traverses its specified arguments. It's assumed that those
locations which also depend on this behaviour do specify it again.

A future implementation may detect the toolchain's capabilities and use
different alternative sets of build instructions. Finding the one syntax
feels like a losing battle. And getting autotools right isn't either the
main goal of this experiment.